pyproject.toml
[tool.poetry]
name = "cruft"
version = "2.4.0"
description = "Allows you to maintain all the necessary cruft for packaging and building projects separate from the code you intentionally write. Built on-top of CookieCutter."
authors = ["Timothy Crosley <timothy.crosley@gmail.com>", "Sambhav Kothari <sambhavs.email@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
cookiecutter = "^1.6"
gitpython = "^3.0"
toml = { version = "^0.10", optional = true }
typer = "^0.3.1"
click = "^7.1.2"
examples = {version = "^1.0.2", optional = true}

[tool.poetry.extras]
pyproject = ["toml"]
examples = ["examples"]

[tool.poetry.dev-dependencies]
vulture = "^1.0"
bandit = "^1.6"
safety = "^1.8"
isort = "^5.3"
flake8-bugbear = "^19.8"
black = {version = "^18.3-alpha.0", allow-prereleases = true}
mypy = "^0.730.0"
ipython = "^7.7"
pytest = "^5.0"
pytest-cov = "^2.7"
pep8-naming = "^0.8.2"
portray = "^1.3.0"
examples = "^1.0.2"
pytest-xdist = "^1.34.0"
pytest-mock = "^3.3.1"

[tool.poetry.scripts]
cruft = "cruft._cli:app"

[tool.portray.mkdocs.theme]
favicon = "art/logo.png"
logo = "art/logo.png"
name = "material"
palette = {primary = "orange", accent = "cyan"}

[tool.cruft]
skip = ["cruft/__init__.py", "tests"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "hug"
